add a handler to rerun a filter on a set of objects or on all alerts between a start and end date, split the api tests in multiple files for readability, add some details about the API in the README
